test(audio): isolate playback factory regression coverage
